+++ previous-versions/files/src/sdk/models/operations/uploadfiles.ts
@@ -3,7 +3,7 @@
  */
 
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
-import * as shared from "../shared";
+import * as shared from "../../../sdk/models/shared";
 import { AxiosResponse } from "axios";
 import { Expose } from "class-transformer";
 
@@ -11,8 +11,8 @@ export class UploadFilesRequestBody extends SpeakeasyBase {
     @SpeakeasyMetadata({ data: "multipart_form, content=true" })
     content: Uint8Array;
 
-    @SpeakeasyMetadata({ data: "multipart_form, name=requestBody" })
-    requestBody: string;
+    @SpeakeasyMetadata({ data: "multipart_form, name=fileName" })
+    fileName: string;
 }
 
 export class UploadFilesRequest extends SpeakeasyBase {